@@ -1,38 +1,40 @@
 <script lang="ts">
   import { unwrapOr } from 'true-myth/maybe';
   import Projects, { type NewProject, type ProjectUpdate } from '$lib/Projects';
-  import Entries from '$lib/Entries';
   import { projects, entries } from '$stores/stores';
   import { empty, isNumber } from '$lib/utils/general';
 
   const { addOrUpdate, getProject, deleteProject } = Projects;
-  const { updateEntries } = Entries;
 
-  export let onSave: (projectId?: UUID) => void;
-  export let projectId: UUID | undefined;
-  export let readOnly: boolean;
+  interface Props {
+    onSave: (projectId?: UUID) => void;
+    projectId: UUID | undefined;
+    readOnly: boolean;
+  }
 
-  $: createOpen = false;
-  $: editOpen = false;
+  let { onSave, projectId, readOnly = $bindable() }: Props = $props();
 
-  let projectCreating: NewProject = { title: undefined, rate: undefined };
-  $: projectCreating;
+  let createOpen = $state(false);
 
-  let selectedProjectId: UUID | undefined;
-  $: selectedProjectId = projectId;
+  let editOpen = $state(false);
 
-  let selectedProject: Project | undefined;
-  $: selectedProject = selectedProjectId
-    ? (unwrapOr(undefined, getProject($projects, selectedProjectId)) as Project)
-    : undefined;
+  const DEFAULT_NEW_PROJECT = { title: undefined, rate: undefined };
 
-  let projectEditing: ProjectUpdate | undefined;
-  $: projectEditing = selectedProject?.id ? { ...selectedProject } : undefined;
+  let projectCreating: NewProject = $state(DEFAULT_NEW_PROJECT);
+
+  let selectedProjectId: UUID | undefined = $state(projectId);
+
+  let selectedProject: Project | undefined = $derived(
+    selectedProjectId ? (unwrapOr(undefined, getProject($projects, selectedProjectId)) as Project) : undefined
+  );
+
+  let projectEditing: ProjectUpdate | undefined = $state();
 
   const isNumberOrEmpty = (val?: string | number) => empty(val) || isNumber(val);
-  $: formValid =
+  let formValid = $derived(
     (createOpen && !empty(projectCreating.title) && isNumberOrEmpty(projectCreating.rate)) ||
-    (editOpen && !empty(projectEditing?.title) && isNumberOrEmpty(projectEditing?.rate));
+      (editOpen && !empty(projectEditing?.title) && isNumberOrEmpty(projectEditing?.rate))
+  );
 
   const byProjectName = (p1: Project, p2: Project) => {
     if (p1.title === p2.title) return 0;
@@ -48,7 +50,7 @@
     }
   };
 
-  const handleSaveClick = (project?: NewProject | ProjectUpdate) => () => {
+  const onSaveClick = (project?: NewProject | ProjectUpdate) => () => {
     if (project === undefined) return;
 
     projects.update((ps) => {
@@ -69,7 +71,7 @@
     editOpen = false;
   };
 
-  $: handleDeleteClick = () => {
+  const onDeleteClick = () => {
     if (selectedProjectId === undefined) return;
 
     const deletedEntries = entries.where((e: TimeEntry) => e.projectId === selectedProjectId);
@@ -86,22 +88,27 @@
 
     // TODO: This component shouldn't need to know about these details. Handle in the models.
     projects.update((ps) => deleteProject(ps, selectedProjectId!));
-    entries.update(
-      (es) =>
-        updateEntries(
-          es,
-          deletedEntries.map((e) => ({ ...e, projectId: undefined }))
-        ).entries
-    );
+
+    entries.updateMany(deletedEntries.map((e) => ({ ...e, projectId: undefined })));
 
     selectedProjectId = undefined;
+  };
+
+  const onEditClick = () => {
+    projectEditing = selectedProject?.id ? { ...selectedProject } : undefined;
+    editOpen = true;
+  };
+
+  const onNewClick = () => {
+    projectCreating = DEFAULT_NEW_PROJECT;
+    createOpen = true;
   };
 </script>
 
 <div class="container">
   {#if readOnly}
     <div class="title-display">
-      <button class="secondary" on:click={() => (readOnly = false)}>
+      <button class="secondary" onclick={() => (readOnly = false)}>
         {selectedProject ? selectedProject.title : 'No project'}
         {selectedProject?.rate ? ` - $${selectedProject?.rate}` : ''}
       </button>
@@ -110,42 +117,42 @@
     <div class="new-project">
       <input
         value={projectEditing.title || ''}
-        on:input={onChange(projectEditing, 'title')}
+        oninput={onChange(projectEditing, 'title')}
         type="text"
         id="edit-title"
         placeholder="Project Title"
       />
       <input
         value={projectEditing.rate || ''}
-        on:input={onChange(projectEditing, 'rate')}
+        oninput={onChange(projectEditing, 'rate')}
         type="text"
         id="edit-rate"
         placeholder="Rate"
       />
       <div class="create-buttons">
-        <button class="secondary" on:click={() => (editOpen = false)}>Cancel</button>
-        <button class="primary" on:click={handleSaveClick(projectEditing)} disabled={!formValid}>Save</button>
+        <button class="secondary" onclick={() => (editOpen = false)}>Cancel</button>
+        <button class="primary" onclick={onSaveClick(projectEditing)} disabled={!formValid}>Save</button>
       </div>
     </div>
   {:else if createOpen}
     <div class="new-project">
       <input
         value={projectCreating.title || ''}
-        on:input={onChange(projectCreating, 'title')}
+        oninput={onChange(projectCreating, 'title')}
         type="text"
         id="title"
         placeholder="Project Title"
       />
       <input
         value={projectCreating.rate || ''}
-        on:input={onChange(projectCreating, 'rate')}
+        oninput={onChange(projectCreating, 'rate')}
         type="text"
         id="rate"
         placeholder="Rate"
       />
       <div class="create-buttons">
-        <button class="secondary" on:click={() => (createOpen = false)}>Cancel</button>
-        <button class="primary" on:click={handleSaveClick(projectCreating)} disabled={!formValid}>Create</button>
+        <button class="secondary" onclick={() => (createOpen = false)}>Cancel</button>
+        <button class="primary" onclick={onSaveClick(projectCreating)} disabled={!formValid}>Create</button>
       </div>
     </div>
   {:else}
@@ -165,18 +172,16 @@
 
     <div class="controls">
       <div>
-        <button on:click={() => (createOpen = true)} class="secondary success">+ New</button>
+        <button onclick={onNewClick} class="secondary success">+ New</button>
         {#if $projects.length > 0}
-          <button on:click={() => (editOpen = true)} class="secondary" disabled={!selectedProjectId}>Edit</button>
-          <button on:click={() => {}} class="error" disabled={!selectedProjectId} on:click={handleDeleteClick}>
-            Delete
-          </button>
+          <button onclick={onEditClick} class="secondary" disabled={!selectedProjectId}>Edit</button>
+          <button onclick={onDeleteClick} class="error" disabled={!selectedProjectId}>Delete</button>
         {/if}
       </div>
 
       <div>
         <button
-          on:click={() => {
+          onclick={() => {
             selectedProjectId = projectId;
             readOnly = true;
             createOpen = false;
@@ -187,7 +192,7 @@
           Cancel
         </button>
         <button
-          on:click={() => {
+          onclick={() => {
             onSave(selectedProjectId);
             readOnly = true;
             createOpen = false;
