@@ -1,5 +1,5 @@
 import Maybe from 'true-myth/maybe';
-import Entries, { type AddEntryResult, type UpdateEntryResult } from '$lib/Entries';
+import Entries, { type AddEntryResult, type UpdateEntryResult, type UpdateEntriesResult } from '$lib/Entries';
 import { createLocalStorageStore } from '$stores/stores';
 
 // An interface for the front end to access long-term storage
@@ -13,6 +13,9 @@ export const entries = {
   },
   update: (entryUpdate: TimeEntryUpdate): UpdateEntryResult => {
     return entriesStorage.updateEntry(entryUpdate);
+  },
+  updateMany: (entryUpdates: TimeEntryUpdate[]): UpdateEntriesResult => {
+    return entriesStorage.updateEntries(entryUpdates);
   },
   add: (newEntry: NewTimeEntry): AddEntryResult => {
     return entriesStorage.addEntry(newEntry);
